feat(client): add user settings dropdown in org creation page
