+++ .changeset/orange-icons-hug.md
@@ -1,0 +1,5 @@
+---
+"@papra/app": patch
+---
+
+Added small header in organization creation page to quickly access the invitations when first organization is being created
