+++ apps/papra-client/src/modules/organizations/pages/create-organization.page.tsx
@@ -1,7 +1,9 @@
 import type { Component } from 'solid-js';
 import { A } from '@solidjs/router';
 import { useI18n } from '@/modules/i18n/i18n.provider';
+import { AppLogo } from '@/modules/ui/components/app-logo';
 import { Button } from '@/modules/ui/components/button';
+import { UserSettingsDropdown } from '@/modules/users/components/user-settings.component';
 import { CreateOrganizationForm } from '../components/create-organization-form.component';
 import { useCreateOrganization } from '../organizations.composables';
 
@@ -11,11 +13,13 @@ export const CreateOrganizationPage: Component = () => {
 
   return (
     <div>
+      <div class="max-w-2xl mx-auto px-6 py-4 flex justify-between items-center border-b">
+        <AppLogo as={A} href="/" />
+
+        <UserSettingsDropdown />
+      </div>
+
       <div class="max-w-md mx-auto pt-12 sm:pt-24 px-6">
-        <Button as={A} href="/" class="mb-4" variant="outline">
-          <div class="i-tabler-arrow-left mr-2" />
-          {t('organizations.create.back')}
-        </Button>
 
         <h1 class="text-xl font-bold">
           {t('organizations.create.title')}
@@ -26,6 +30,10 @@ export const CreateOrganizationPage: Component = () => {
         </p>
 
         <CreateOrganizationForm onSubmit={createOrganization} />
+        <Button as={A} href="/" class="mb-4 w-full" variant="outline">
+          <div class="i-tabler-arrow-left mr-2" />
+          {t('organizations.create.back')}
+        </Button>
       </div>
     </div>
   );
